docs: add CAL catalog
